需求[#21996]:SR-20201214-02-19 任务超期检查接口
    [描述说明]: 任务超期检查接口
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/task/entity/TaskOnExecuteInfo.java
@@ -35,7 +35,7 @@ public class TaskOnExecuteInfo {
     private String cruiseTypeName;
 
     @ApiModelProperty(value = "任务进度" )
-    private Float taskProgress;
+    private Integer taskProgress;
 
     @ApiModelProperty(value = "异常结果个数" )
     private Integer alarmCount;
